readability mostly
went through the code, writing comments to get a better grasp of the entirety of it. also change the searchterm filter to look for data that "includes" the search term instead of just "startsWith" for more flexible searching. also moved the getObjects function to the bottom om the file for more readability