Search.jsx component
created a component that has a searchbar with autocomplete dropdown. have managed to apply it to both the monsters tab and the spells tab so i know it is modular and can be used other places where i might need it. also cleaned out some old code that i dont need anymore